Changed duplicate id handling, static objids
- I noticed when targetting a client-spawned object (EK brazier), it
sent a target packet to the client with all 8 bytes filled, instead of
just the 4 stated in the buildout.  It seems the same 4 digits get added
to all buildout ids.  As a result, all buildout int-objectIds now have
the follow bytes added to them: FF FF 86 F9

Not sure about snapshot objects, but I imagine it's the same for them.

- I had to allow duplicate objects to be spawned in order for heroics
and instances to work.  It gives every duplicate object a new objectId
and also replaces the parentIds as well by storing the changes in a map
on each buildout area cycle.