Player flag is set when you align to a faction ->
When you join a faction, #16 is set to true.

When you go combatant, #32 is set to true.  The resulting flag: 48

When you go SF, #1, #2 and #4 are set to true.  The resulting flag: 55

When you go back to combatant, #1, #2 and #4 are set to false.  The resulting flag: 48

The error was either in the Player flag not being set anywhere, or setPvpStatus setting it instead of doing OR/XOR operations.

(The result of this would be bit #16 wouldn't be set to true, meaning SF would be 39 instead of 55)

We could still set all of the flags to true for SF without harm for safety, it just wasn't theoretically necessary.

Additionally any players that login automatically were getting #16 sent to the client by sendBaselines() regardless of if they had a faction, which should now be fixed.