Buffs refactored to use DatatableVisitor on server load
Originally, everytime a new Buff instance was created, it would loop
through the Datatable Visitor everytime. Now it's only done once on
server startup and put into a buffMap. A copy of the base buff is then
done through the constructor of Buff class should a new instance of that
buff type needs to be created from the one inside the map.