Fixed some housing bugs and Tusken Raider mobile template - see extended
Fixed a housing issue where the structure management terminal would be
returned in the item list and could be deleted via the delete all items
radial.

Fixed item list to display proper names of objects.

Fixed the Tusken Raider mobile template to display the proper creature
name and changed their default weapons from an E11 and a T21 to a Tusken
Rifle.