Temporarily Fixes #442: Position secured cant be removed
The problem related to doSelfBuff always calling and thinking that the
position secured buff wanted to be added all the time, which isn't the
case. Right now I added a hasBuff check to doSelfBuff for the position
secured buff which will remove the buff. I didn't want to flat out just
remove a buff no matter what command was called because that would
affect any buffs that wanted to be refreshed and not removed.