Removed getObjectByFirstName from ChatService (deprecated), Changed getting object by names methods to account for all type cases
These changes should fix a lot of issues (mainly chat related like
mails, tells, chat channels) for anything that was using the out-dated
getObjectByFirstName() in chatservice. Also it'll allow us to not have
to go through uppercase and lowercase versions every time we try and
pull an object by names.