Fixed some long-term buff issues.
We've had a completely incorrect buff structure in baselines for the last year due to some small misinterpretations of its structure.  It worked because there was the right amount of bytes, but otherwise the structure wasn't right and variables have been in the wrong place.  Specifically, we interpreted the stack variable (which comes last) to be the beginning of all buffs except the first, assumed there was an abnormal starting buff, and treated the final stack variable as an unknown variable that comes after the buff list.  This is most likely also why it didn't work without this "default buff".  The structure has been correct in Buff.getBytes() and buff deltas for a while but has been incorrect for baselines.