Added hate to AI (distinct from damage).
Hate is not the same as damage.

AI is meant to choose their target based on who's most hated, not who's
dealt the most damage.

Hate is affected by the hate fields in combat commands, not necessarily
the damage of those commands.

An example is the Guardian Strike ability which deals zero damage but
increases hate greatly.

There is still a purpose for HighestDamageDealer in locations such as
DeathState where the person who gets the loot is the highest damage
dealer.  Damage dealt is also used in XP calculation.